+++ test/test_thd_distributed.py
@@ -11,7 +11,7 @@ from functools import reduce, wraps
 
 import torch
 import torch.cuda
-import torch.distributed as dist
+import torch.distributed.deprecated as dist
 import torch.nn as nn
 import torch.nn.functional as F
 import torch.optim as optim
@@ -985,7 +985,7 @@ class _DistTestBase(object):
         # DDP training setup
         model_DDP = copy.deepcopy(model)
         model_DDP.cuda(gpu_subset[0])
-        model_DDP = nn.parallel.DistributedDataParallel(
+        model_DDP = nn.parallel.deprecated.DistributedDataParallel(
             model_DDP, device_ids=gpu_subset
         )
 
@@ -1019,7 +1019,7 @@ class _DistTestBase(object):
 
         # DDP-CPU training setup
         model_DDP = copy.deepcopy(model_base)
-        model_DDP = nn.parallel.DistributedDataParallelCPU(model_DDP)
+        model_DDP = nn.parallel.deprecated.DistributedDataParallelCPU(model_DDP)
 
         # dummy data initialization
         local_bs = 2
