+++ torch/nn/parallel/__init__.py
@@ -4,8 +4,7 @@ from .data_parallel import DataParallel, data_parallel
 from .scatter_gather import scatter, gather
 from .distributed import DistributedDataParallel
 from .distributed_cpu import DistributedDataParallelCPU
-from .distributed_c10d import _DistributedDataParallelC10d
-from .distributed_c10d_cpu import _DistributedDataParallelC10dCPU
+import torch.nn.parallel.deprecated
 
 __all__ = ['replicate', 'scatter', 'parallel_apply', 'gather', 'data_parallel',
            'DataParallel', 'DistributedDataParallel', 'DistributedDataParallelCPU']
