+++ torch/nn/parallel/deprecated/__init__.py
@@ -1,0 +1,4 @@
+from .distributed import DistributedDataParallel
+from .distributed_cpu import DistributedDataParallelCPU
+
+__all__ = ['DistributedDataParallel', 'DistributedDataParallelCPU']
