+++ torch/nn/parallel/distributed_cpu.py
@@ -7,9 +7,10 @@ from torch.autograd import Variable
 
 
 class DistributedDataParallelCPU(Module):
-    r"""Implements distributed data parallelism for CPU at the module level.
+    r"""Implements distributed data parallelism for CPU at the module level
+    and it will use PyTorch's new distributed package: c10d.
 
-    This module support the ``mpi``, ``gloo``, ``tcp`` backends.
+    This module support the ``mpi``, ``gloo``, backends.
 
     This container parallelizes the application of the given module by
     splitting the input across the specified devices by chunking in the batch
